Consolidate checkout E2E helpers
