+++ e2e/tests/mobile/registered-shopper.spec.js
@@ -13,9 +13,9 @@ const {
     validateWishlist,
     loginShopper,
     navigateToPDPMobile,
-    answerConsentTrackingForm
+    answerConsentTrackingForm,
+    advanceToPayment
 } = require('../../scripts/pageHelpers')
-const {advanceToPayment} = require('../../scripts/checkout')
 const {generateUserCredentials, getCreditCardExpiry} = require('../../scripts/utils.js')
 const {clearCartAndWishlist} = require('../../scripts/cleanup.js')
